fix: SAML Connection response type and accept pagination params
This commit fixes the SAML Connection ListAll() method, to return the proper data (paginated) and to also accept any pagination params that the user might want to use
